Make sure all files are protected by build tags
