+++ fsevents_test.go
@@ -56,7 +56,7 @@ func TestBasicExample(t *testing.T) {
 	if err != nil {
 		t.Fatal(err)
 	}
-	select{
+	select {
 	case <-wait:
 	case <-time.After(5 * time.Second):
 		t.Fatal("timed out waiting for event")
@@ -147,5 +147,27 @@ func TestIssue48(t *testing.T) {
 	if err == nil {
 		es.Stop()
 		t.Fatal("eventstream error was not detected on >4096 files in watchlist")
+	}
+}
+
+func TestRegistry(t *testing.T) {
+	if registry.m == nil {
+		t.Fatal("registry not initialized at start")
+	}
+
+	es := &EventStream{}
+	i := registry.Add(es)
+
+	if registry.Get(i) == nil {
+		t.Fatal("failed to retrieve es from registry")
+	}
+
+	if es != registry.Get(i) {
+		t.Errorf("eventstream did not match what was found in the registry")
+	}
+
+	registry.Delete(i)
+	if registry.Get(i) != nil {
+		t.Error("failed to delete registry")
 	}
 }
